\name{breakpoints}
\alias{breakpoints}
\docType{data}
\title{
    Putative breakpoints of chimeric reads
}
\description{
    This example holds two consensus (pathogenic and reciproce) breakpoints of 12 chimeric reads indicating an inversion on chromosome 16.
    The Breakpoints object gives access to the breakpoint locations as well as alignment information for each of the 12 reads.
}
\usage{data(breakpoints)}

\format{
    Formal class 'Breakpoints'
}

\source{
    \sQuote{Targeted next-generation sequencing detects point mutations,
    insertions, deletions, and balanced chromosomal rearrangements as
    well as identifies novel leukemia-specific fusion genes in a single
    procedure} (Leukemia, submitted)
}

\examples{
    data(breakpoints)
}
\keyword{datasets}
