\name{captureArray}
\alias{captureArray}
\docType{data}
\title{Custom capture array design}
\description{
Design of a custom Roche NimbleGen 385k capture array. The array captures
short segments corresponding to all exon regions of 92 distinct 
target genes (genome build hg19). In addition, contiguous genomic
regions were represented for three additional genes, i.e. CBFB, MLL, and RUNX1.
}
\usage{data(captureArray)}
\format{
Formal class 'CompressedIRangesList'
}
\source{
\sQuote{Targeted next-generation sequencing detects point mutations, 
insertions, deletions, and balanced chromosomal rearrangements as 
well as identifies novel leukemia-specific fusion genes in a single
procedure} (Leukemia, submitted)
}
\examples{
data(captureArray)
}
\keyword{datasets}
