\name{mapperSetExample}
\alias{mapperSetExample}
\docType{data}
\title{
    GS Reference Mapper data import
}

\description{
    This is an example of an \code{link{MapperSet-class}} object containing the output of Roche's GS Reference Mapper Software.
    It consists of 3 samples and 111 variants.
}

\usage{data(mapperSetExample)}

\format{
    Formal class 'MapperSet'
}

\source{
    \sQuote{Targeted next-generation sequencing detects point mutations,
    insertions, deletions, and balanced chromosomal rearrangements as
    well as identifies novel leukemia-specific fusion genes in a single
    procedure} (Leukemia, submitted)
}

\examples{
    data(mapperSetExample)
    mapperSetExample
}
\keyword{datasets}
