\name{mutationInfo}
\alias{mutationInfo}
\docType{data}
\title{
	Example data for \code{\link{plotVariants}}
}
\description{
 	This data.frame is part of the vignette example of the \code{\link{plotVariants}} function. It contains annotations for the different
  	mutations occurring in the example data. It has columns "mutation", "legend" and "color".
}

\usage{data(plotVariantsExample)}

\format{
	data.frame
}

\examples{
    data(plotVariantsExample)
    mutationInfo
}
\keyword{datasets}
