\name{regions}
\alias{regions}
\docType{data}

\title{
        Example data for \code{\link{plotVariants}}
}
\description{
        This data.frame is part of the vignette example of the \code{\link{plotVariants}} function. It has the columns "name", "start", "end" and
  	"color". The plot will highlight these regions with the given colors and print their name in the legend.
}

\usage{data(plotVariantsExample)}

\format{
	data.frame
}

\examples{
   	data(plotVariantsExample)
	regions
}
\keyword{datasets}
