\name{variants}
\alias{variants}
\docType{data}

\title{
        Example data for \code{\link{plotVariants}}
}
\description{
        This data.frame is part of the vignette example of the \code{\link{plotVariants}} function. It has the columns
  	"label", "pos" "mutation" and "color" specifying an annotation for each mutation, its position, the mutation type and an individual color.
	The position is given as amino acids / codons.
}

\usage{data(plotVariantsExample)}

\format{
	data.frame
}

\examples{
    	data(plotVariantsExample)
    	variants
}
\keyword{datasets}
